add wrapper function to test actions
