+++ packages/python/tests/run.sh
@@ -4,7 +4,7 @@ set -euo pipefail
 export DEBUG_HTTP_INCLUDE_BODY=1
 
 SINGLE=true
-SINGLE=false
+# SINGLE=false
 
 if [ "$SINGLE" = 'false' ]; then
   coverage run --source packages/python/m -m pytest -p no:logging
@@ -12,7 +12,7 @@ if [ "$SINGLE" = 'false' ]; then
 else
   # To run specific tests:
   source="packages/python/m/github/actions"
-  filter="test_testing"
+  filter="test_github_actions"
   coverage run --source "$source" -m pytest -p no:logging packages/python -vv -k "$filter"
   coverage report -m
 fi
